+++ serialize/SerializeElement.h
@@ -41,10 +41,11 @@ public:
 		return SerializeAdapter<T>::getSerializedSize(&entry);
 	}
 
-	virtual ReturnValue_t deSerialize(const uint8_t** buffer, int32_t* size,
+	virtual ReturnValue_t deSerialize(const uint8_t** buffer, ssize_t* size,
 			bool bigEndian) {
 		return SerializeAdapter<T>::deSerialize(&entry, buffer, size, bigEndian);
 	}
+
 	operator T() {
 		return entry;
 	}
